+++ packages/extensions/core/src/lib/pipeline/plugins/enum-deduplication/index.ts
@@ -1,0 +1,2 @@
+export * from "./enum-deduplication-plugin";
+export * from "./enum-deduplicator";