+++ packages/libs/datastore/src/data-store/data-store.ts
@@ -145,12 +145,17 @@ class ReadThroughDataSource extends DataSource {
         // populate cache
         let data: string | null = null;
         try {
-          data = (await this.fs.ReadFile(uri)) || (await ReadUri(uri));
+          data = (await this.fs.read(uri)) || (await ReadUri(uri));
           if (data) {
             const parent = ParentFolderUri(uri) || "";
             // hack to let $(this-folder) resolve to the location...
             data = data.replace(/\$\(this-folder\)\/*/g, parent);
           }
+        } catch (e) {
+          // TODO-TIM: Reeenable this log with new logging system https://github.com/Azure/autorest/issues/3988
+          // Disabled this as it creates too much noise for some expected failure(Cannot find samples)
+          // eslint-disable-next-line no-console
+          // console.error("Unexpected error trying to read file", e);
         } finally {
           if (!data) {
             // eslint-disable-next-line no-unsafe-finally