+++ packages/libs/datastore/src/main.ts
@@ -3,9 +3,9 @@
  *  Licensed under the MIT License. See License.txt in the project root for license information.
  *--------------------------------------------------------------------------------------------*/
 // export * from "./proxy"
-export * from "./file-system/file-system";
+export * from "./file-system";
 export * from "./json-path/json-path";
-export * from "./json-pointer/json-pointer";
+export * from "./json-pointer";
 export * from "./graph-builder";
 export * from "./data-store/data-store";
 export * from "./cancellation";
@@ -19,5 +19,6 @@ export * from "./parsing/yaml";
 export * from "./source-map/blaming";
 export * from "./parsing/stable-object";
 export * from "./processor";
+export * from "./transformer-via-pointer";
 
 export { safeEval, createSandbox } from "@azure-tools/codegen";