+++ packages/libs/oai2-to-oai3/test/scenarios/scenarios.test.ts
@@ -54,4 +54,9 @@ describe("Scenario testings", () => {
     // The expected result is the parmaeter to be included/expanded in the OpenAPI3 server property.
     await expectInputsMatchSnapshots("parameterized-host-parameters", ["swagger.json"]);
   });
+
+  fit("Convert enums using $ref object as values", async () => {
+    // The expected result is the $ref in `enum` has been updated to the openapi 3 format.
+    await expectInputsMatchSnapshots("enums", ["swagger.json"]);
+  });
 });