+++ packages/libs/codegen/CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log - @azure-tools/codegen
 
-This log was last generated on Thu, 04 Feb 2021 19:16:37 GMT and should not be manually modified.
+This log was last generated on Tue, 16 Mar 2021 15:52:56 GMT and should not be manually modified.
+
+## 2.5.294
+Tue, 16 Mar 2021 15:52:56 GMT
+
+### Patches
+
+- Bump dependencies versions
 
 ## 2.5.293
 Thu, 04 Feb 2021 19:05:18 GMT